emp/requirements.txt
yagmail
slack_sdk
python-dotenv
pandas
openpyxl
